LBJ-208 moving some Classes that are not meant to be instantiated or used directly by library users to the impl packages
