+++ src/main/java/org/zalando/nakadiproducer/NakadiProducerAutoConfiguration.java
@@ -25,10 +25,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.zalando.fahrschein.NakadiClient;
-import org.zalando.nakadiproducer.snapshots.SnapshotEventCreationEndpoint;
-import org.zalando.nakadiproducer.snapshots.SnapshotEventCreationMvcEndpoint;
+import org.zalando.nakadiproducer.snapshots.impl.SnapshotEventCreationEndpoint;
+import org.zalando.nakadiproducer.snapshots.impl.SnapshotEventCreationMvcEndpoint;
 import org.zalando.nakadiproducer.snapshots.SnapshotEventProvider;
-import org.zalando.nakadiproducer.snapshots.SnapshotEventProviderNotImplementedException;
+import org.zalando.nakadiproducer.snapshots.impl.SnapshotEventProviderNotImplementedException;
 import org.zalando.nakadiproducer.snapshots.impl.SnapshotCreationService;
 import org.zalando.stups.tokens.Tokens;
 import org.zalando.tracer.Tracer;
